refactor(crawler): move models to nsysu_crawler package
