+++ lib/api/selcrs_helper.dart
@@ -8,9 +8,7 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/foundation.dart';
 import 'package:html/dom.dart' as dom;
 import 'package:html/parser.dart';
-import 'package:nsysu_ap/models/options.dart';
-import 'package:nsysu_ap/models/pre_score.dart';
-import 'package:nsysu_ap/models/score_semester_data.dart';
+import 'package:nsysu_crawler/nsysu_crawler.dart';
 import 'package:nsysu_ap/utils/big5/big5.dart';
 import 'package:nsysu_ap/utils/utils.dart';
 import 'package:sprintf/sprintf.dart';
