+++ lib/api/tuition_helper.dart
@@ -7,7 +7,7 @@ import 'package:flutter/foundation.dart';
 import 'package:dio_cookie_manager/dio_cookie_manager.dart';
 import 'package:html/dom.dart';
 import 'package:html/parser.dart';
-import 'package:nsysu_ap/models/tuition_and_fees.dart';
+import 'package:nsysu_crawler/nsysu_crawler.dart';
 
 class TuitionHelper {
   static const String basePATH = 'https://tfstu.nsysu.edu.tw';
