+++ lib/pages/bus/bus_time_page.dart
@@ -3,8 +3,7 @@ import 'dart:async';
 import 'package:ap_common/ap_common.dart';
 import 'package:flutter/material.dart';
 import 'package:nsysu_ap/api/bus_helper.dart';
-import 'package:nsysu_ap/models/bus_info.dart';
-import 'package:nsysu_ap/models/bus_time.dart';
+import 'package:nsysu_crawler/nsysu_crawler.dart';
 import 'package:nsysu_ap/utils/app_localizations.dart';
 
 class BusTimePage extends StatefulWidget {
