+++ lib/pages/tuition_and_fees_page.dart
@@ -6,8 +6,9 @@ import 'package:ap_common/ap_common.dart';
 import 'package:flutter/material.dart';
 import 'package:nsysu_ap/api/selcrs_helper.dart';
 import 'package:nsysu_ap/api/tuition_helper.dart';
-import 'package:nsysu_ap/models/tuition_and_fees.dart';
+import 'package:nsysu_ap/extensions/tuition_and_fees_ui_extension.dart';
 import 'package:nsysu_ap/utils/app_localizations.dart';
+import 'package:nsysu_crawler/nsysu_crawler.dart';
 
 class TuitionAndFeesPage extends StatefulWidget {
   const TuitionAndFeesPage({super.key});
