+++ packages/nsysu_crawler/lib/src/models/bus_time.dart
