+++ packages/nsysu_crawler/lib/src/models/graduation_report_data.dart
