+++ packages/nsysu_crawler/lib/src/models/pre_score.dart
