+++ packages/nsysu_crawler/lib/src/models/tuition_and_fees.dart
@@ -1,0 +1,19 @@
+class TuitionAndFees {
+  final String titleZH;
+  final String titleEN;
+  final String amount;
+  final String paymentStatusZH;
+  final String paymentStatusEN;
+  final String dateOfPayment;
+  final String serialNumber;
+
+  TuitionAndFees({
+    required this.titleZH,
+    required this.titleEN,
+    required this.amount,
+    required this.paymentStatusZH,
+    required this.paymentStatusEN,
+    required this.dateOfPayment,
+    required this.serialNumber,
+  });
+}
